pyproject.toml
[project]
name = "events-api"
version = "1.0.0"
requires-python = ">=3.13"
readme = "README.md"
dependencies = [
    "fastapi[standard]==0.135.1",
    "pydantic[email]",
    "sqlalchemy[asyncio]==2.0.48",
    "asyncpg==0.31.0",
    "alembic==1.18.4",
    "pyjwt==2.12.1",
    "bcrypt==5.0.0",
    "psycopg2-binary==2.9.11",
]

[tool.uv]
package = false

[dependency-groups]
test = [
    "pytest==9.0.2",
    "pytest-cov (==7.1.0)",
    "aiosqlite==0.22.1",
    "pytest-asyncio==1.3.0",
]
dev = [
    "pre-commit (==4.5.1)",
    "ruff (==0.15.7)",
]

# Ruff configuration
[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint]
select = [
    "E", # pycodestyle
    "F", # Pyflakes
    "UP", # pyupgrade
    "B", # flake8-bugbear
    "SIM", # flake8-simplify
    "I", # isort
]
ignore = ["E203", "E305", "B008"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
